web/requirements.txt
Babel==2.3.4
Django==1.10.4
Fabric==1.13.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.2
Pygments==2.1.3
Sphinx==1.5.1
alabaster==0.7.9
argon2-cffi==16.3.0
argparse==1.2.1
bcrypt==3.1.1
blessings==1.6
bpython==0.16
cffi==1.9.1
configparser==3.5.0
coverage==4.2
cryptography==1.7.1
curtsies==0.2.11
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.6
django-tagging==0.4.5
django-taggit==0.21.3
docutils==0.13.1
enum34==1.1.6
flake8==3.2.1
greenlet==0.4.11
idna==2.1
imagesize==0.7.1
ipaddress==1.0.17
mccabe==0.5.3
ndg-httpsclient==0.4.2
numpy==1.11.2
paramiko==2.1.1
pluggy==0.4.0
psycopg2==2.6.2
py==1.4.32
pyOpenSSL==16.2.0
pyasn1==0.1.9
pycodestyle==2.2.0
pycparser==2.17
pyflakes==1.3.0
pytest==3.0.5
pytz==2016.10
rcssmin==1.0.6
requests==2.12.4
rjsmin==1.0.12
six==1.10.0
snowballstemmer==1.2.1
sqlparse==0.2.2
tox==2.5.0
virtualenv==15.1.0
wcwidth==0.1.7
wsgiref==0.1.2
